[NO-ISSUE] 동영상 업로드 관련 코드 리팩토링 (#30)
* [NO-ISSUE] chore: yml 파일 청소

* [NO-ISSUE] chore: local, dev 환경으로 제한

* [NO-ISSUE] test: @SpringBootTest recover

* [NO-ISSUE] chore: 사용하지 않는 파일 제거

* [NO-ISSUE] refactor: 패키지 변경 (utils -> util)

* [NO-ISSUE] refactor: 상수 추출 및 컨벤션 적용

* [NO-ISSUE] chore: jdk distribution 변경

* [NO-ISSUE] test: System.out.println 추가

* [NO-ISSUE] chore: test 환경 log level debug 적용

* [NO-ISSUE] chore: jdk distribution adopt로 복구
