+++ .github/workflows/ON_SPRINT.yml
@@ -59,4 +59,3 @@ jobs:
             cd /app/longs-env/env-dev
             docker image rm --force vidigummy/svr
             docker compose up -d --force-recreate
-
