+++ src/main/java/com/sds/actlongs/config/FFmpegConfig.java
@@ -5,12 +5,15 @@ import java.io.IOException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 import net.bramp.ffmpeg.FFmpeg;
 import net.bramp.ffmpeg.FFprobe;
 
+@Profile({"local", "dev"})
 @Configuration
 public class FFmpegConfig {
+
 	@Value("${ffmpeg.path}")
 	private String ffmpegPath;
 
@@ -26,4 +29,5 @@ public class FFmpegConfig {
 	public FFprobe ffProbe() throws IOException {
 		return new FFprobe(ffprobePath);
 	}
+
 }
