+++ src/main/java/com/sds/actlongs/controller/upload/UploadController.java
@@ -3,6 +3,7 @@ package com.sds.actlongs.controller.upload;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -15,6 +16,7 @@ import lombok.RequiredArgsConstructor;
 import com.sds.actlongs.service.upload.UploadService;
 import com.sds.actlongs.service.upload.dto.UploadResponseDto;
 
+@Profile({"local", "dev"})
 @Api(tags = "동영상 업로드 Api")
 @RestController
 @RequiredArgsConstructor
