+++ src/main/java/com/sds/actlongs/domain/video/repository/VideoRepository.java
@@ -1,5 +1,9 @@
 package com.sds.actlongs.domain.video.repository;
 
-public interface VideoRepository {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.sds.actlongs.domain.video.entity.Video;
+
+public interface VideoRepository extends JpaRepository<Video, Long> {
 
 }
