+++ src/main/java/com/sds/actlongs/service/upload/TusUploadService.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
@@ -13,10 +14,11 @@ import me.desair.tus.server.exception.TusException;
 import me.desair.tus.server.upload.UploadInfo;
 
 import com.sds.actlongs.service.upload.dto.UploadResponseDto;
-import com.sds.actlongs.utils.manage.file.FileManage;
-import com.sds.actlongs.utils.thumbnail.ThumbnailExtractor;
-import com.sds.actlongs.utils.uuid.UuidGenerate;
+import com.sds.actlongs.util.manage.file.FileManage;
+import com.sds.actlongs.util.thumbnail.ThumbnailExtractor;
+import com.sds.actlongs.util.uuid.UuidGenerate;
 
+@Profile({"local", "dev"})
 @Service
 @RequiredArgsConstructor
 public class TusUploadService implements UploadService {
