+++ src/main/java/com/sds/actlongs/util/Constants.java
@@ -8,6 +8,7 @@ public final class Constants {
 
 	public static final String EMPTY = "";
 	public static final String SPACE = " ";
+	public static final String HYPHEN = "-";
 	public static final String DOT = ".";
 	public static final String WILDCARD = "*";
 	public static final String ALL_PATH = "/**";
