+++ src/main/java/com/sds/actlongs/util/duration/DurationExtractor.java
@@ -1,0 +1,8 @@
+package com.sds.actlongs.util.duration;
+
+import java.io.File;
+import java.io.IOException;
+
+public interface DurationExtractor {
+	Double extract(String sourcePath);
+}
