+++ src/main/java/com/sds/actlongs/util/thumbnail/ThumbnailExtractor.java
@@ -1,0 +1,11 @@
+package com.sds.actlongs.util.thumbnail;
+
+import java.io.File;
+import java.io.IOException;
+
+public interface ThumbnailExtractor {
+
+	String extract(String fileName);
+
+	String generateDefaultThumbnail(String fileName);
+}
