+++ src/main/java/com/sds/actlongs/util/uuid/UuidGenerate.java
@@ -1,0 +1,5 @@
+package com.sds.actlongs.util.uuid;
+
+public interface UuidGenerate {
+	String getVodUuid();
+}
