--- src/main/java/com/sds/actlongs/utils/duration/DurationExtractor.java
@@ -1,8 +1,0 @@
-package com.sds.actlongs.utils.duration;
-
-import java.io.File;
-import java.io.IOException;
-
-public interface DurationExtractor {
-	Double extract(String sourcePath);
-}
