--- src/main/java/com/sds/actlongs/utils/uuid/UuidGenerate.java
@@ -1,5 +1,0 @@
-package com.sds.actlongs.utils.uuid;
-
-public interface UuidGenerate {
-	String getVodUuid();
-}
