--- src/main/java/com/sds/actlongs/utils/uuid/UuidGenerateImpl.java
@@ -1,15 +1,0 @@
-package com.sds.actlongs.utils.uuid;
-
-import java.util.UUID;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class UuidGenerateImpl implements UuidGenerate {
-
-	@Override
-	public String getVodUuid() {
-		return UUID.randomUUID().toString().replaceAll("-", "");
-	}
-
-}
