+++ src/test/java/com/sds/actlongs/ActLongsApplicationTests.java
@@ -3,9 +3,12 @@ package com.sds.actlongs;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
+@SpringBootTest
 class ActLongsApplicationTests {
 
+	@Test
 	void contextLoads() {
 
 	}
+
 }
